Updated string copy functions
Currently, the char data is copied using standard library functions which has the limitation of not copying the data beyond 1st NULL char.

Per RFC 3261, Binary Data is allowed to be added as part of SIP Event Body.
In such case, the 1st occurrence of binary '0' causes the problem.

Updating the string Copy function here to solve this issue